refactor: remove version-specific jest adapters
the jest-test-adapters are functionally equivalent now so we can refactor them
into one.
