+++ packages/jest-runner/src/jest-test-adapters/index.ts
@@ -1,2 +1,2 @@
-export type { JestTestAdapter } from './jest-test-adapter.js';
+export { JestTestAdapter } from './jest-test-adapter.js';
 export { jestTestAdapterFactory } from './jest-test-adapter-factory.js';
